# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "O Chefia",
        "description": "Backend para um sistema de PDV e controle simples empresarial.",
        "contact": {
            "email": "contato.wanjalagus@outlook.com.br"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/license/mit/"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/",
            "description": "O Chefia servidor"
        }
    ],
    "paths": {
        "/api/v1/autenticacao/login": {
            "post": {
                "tags": [
                    "Login de funcionários e clientes"
                ],
                "operationId": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "joe_doe@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "5|FggYlCsXzqKISnrGA6NHZzJtQaRD2YGMZQ1yDPvf"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int64",
                                                    "example": "10"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Joe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "joe_doe@email.com"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "empresa_token": {
                                            "type": "string",
                                            "example": "eyJlbXByZXNhX2lkIjo1LCJlbXByZXNhX2NucGoiOiI0NjUxOTQyMzAwMDExMyJ9"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Ao tentar entrar no sistema com uma senha que não é a cadastrada com o email enviado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "example": "A senha é inválida."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ao tentar entrar no sistema com um email que não existe",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "erro": {
                                            "example": "O email inserido não está cadastrado, insira um válido ou se cadastre no sistema."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}